+++ src/commands/dev.ts
@@ -130,6 +130,13 @@ export default class Dev extends BaseCommand {
         char: 'd',
       }),
     },
+    debug: {
+      non_sensitive: true,
+      ...Flags.string({
+        description: `[default: true] Turn debug mode on (true) or off (false)`,
+        default: 'true',
+      }),
+    },
     arg: Flags.string({
       description: 'Build arg(s) to pass to docker build',
       multiple: true,
@@ -395,8 +402,9 @@ export default class Dev extends BaseCommand {
 
     const component_options: ComponentConfigOpts = { map_all_interfaces: !flags.production && !duplicates, interfaces: interfaces_map };
 
+    const debug = flags.debug === 'true';
     for (const component_version of component_versions) {
-      const component_config = await dependency_manager.loadComponentSpec(component_version, component_options);
+      const component_config = await dependency_manager.loadComponentSpec(component_version, component_options, debug);
 
       if (flags.recursive) {
         const dependency_configs = await dependency_manager.loadComponentSpecs(component_config.metadata.ref);