+++ src/common/docker-compose/project.ts
@@ -1,0 +1,5 @@
+export type DockerComposeProject = {
+  Name: string;
+  Status: string;
+  ConfigFiles: string;
+};