+++ src/dependency-manager/spec/component-spec.ts
@@ -18,12 +18,12 @@ export interface ComponentInstanceMetadata {
   readonly instance_id?: string;
   readonly instance_date: Date;
 
-  proxy_port_mapping?: Dictionary<number>;
-
   file?: {
     path: string;
     contents: string;
   }
+
+  interpolated?: boolean;
 }
 
 @JSONSchema({