+++ test/dependency-manager/volumes.test.ts
@@ -51,7 +51,7 @@ describe('volumes spec v1', () => {
       'test/component': '/component/component.yml'
     });
     const graph = await manager.getGraph([
-      await manager.loadComponentSpec('test/component')
+      await manager.loadComponentSpec('test/component', {}, true)
     ]);
     const template = await DockerComposeUtils.generate(graph);
     expect(template.services[test_component_api_safe_ref].volumes).has.members([`${path.resolve('/component/data')}:/data`])
@@ -114,7 +114,7 @@ describe('volumes spec v1', () => {
       'test/component': '/component/component.yml'
     });
     const graph = await manager.getGraph([
-      await manager.loadComponentSpec('test/component')
+      await manager.loadComponentSpec('test/component', {}, true)
     ])
     const template = await DockerComposeUtils.generate(graph);
     expect(template.services[test_component_api_safe_ref].volumes).has.members(['api-data:/data', 'api-data2:/data2', `${path.resolve('/component/data3')}:/data3`])